+++ ALGORITHMS/Sorting/Selection Sort/selctionSort.py
